backend/requirements.txt
Django==1.8.7
djangorestframework==3.3.1
django-cors-headers==1.1.0
